@@ -10,26 +10,36 @@
   },
   "dependencies": {
     "@azure/msal-browser": "^2.20.0",
-    "vue": "^2.6.14",
-    "vuetify": "^2.6.2",
-    "vuex": "^3.6.2"
+    "pinia": "^2.0.17",
+    "vue": "^2.7.8",
+    "vue-class-component": "^7.2.3",
+    "vue-property-decorator": "^9.1.2",
+    "vuetify": "^2.6.8"
   },
   "devDependencies": {
     "@babel/core": "^7.12.16",
     "@babel/eslint-parser": "^7.12.16",
     "@mdi/js": "^6.5.95",
+    "@pinia/testing": "^0.0.13",
+    "@types/jest": "^27.0.1",
+    "@typescript-eslint/eslint-plugin": "^5.4.0",
+    "@typescript-eslint/parser": "^5.4.0",
     "@vue/cli-plugin-babel": "~5.0.0-rc.1",
     "@vue/cli-plugin-eslint": "~5.0.0-rc.1",
+    "@vue/cli-plugin-typescript": "~5.0.0",
     "@vue/cli-plugin-unit-jest": "~5.0.0",
     "@vue/cli-service": "~5.0.0-rc.1",
-    "@vue/test-utils": "^1.1.3",
+    "@vue/eslint-config-typescript": "^11.0.0",
+    "@vue/test-utils": "^1.3.0",
     "@vue/vue2-jest": "^27.0.0-alpha.2",
+    "eslint-plugin-vue": "^9.4.0",
     "babel-jest": "^27.0.6",
     "eslint": "^7.32.0",
-    "eslint-plugin-vue": "^9.3.0",
     "jest": "^27.0.5",
     "sass": "^1.43.2",
     "sass-loader": "^10.0.0",
+    "ts-jest": "^27.0.4",
+    "typescript": "~4.5.5",
     "vue-cli-plugin-vuetify": "~2.4.5",
     "vue-template-compiler": "^2.6.11",
     "vuetify-loader": "^1.7.0",
@@ -42,10 +52,11 @@
     },
     "extends": [
       "plugin:vue/essential",
-      "eslint:recommended"
+      "eslint:recommended",
+      "@vue/typescript"
     ],
     "parserOptions": {
-      "parser": "@babel/eslint-parser"
+      "parser": "@typescript-eslint/parser"
     },
     "rules": {},
     "overrides": [
