@@ -1,17 +1,11 @@
 <template>
   <v-col>
-    <v-menu
-      v-if="$store.state.account"
-      bottom
-      min-width="300px"
-      rounded
-      offset-y
-    >
+    <v-menu v-if="user.account" bottom min-width="300px" rounded offset-y>
       <template v-slot:activator="{ on }">
         <v-btn icon x-large v-on="on">
           <v-avatar color="indigo" size="48">
             <span class="white--text text-h6">{{
-              $store.state.account.initials
+              getInitials(user.account)
             }}</span>
           </v-avatar>
         </v-btn>
@@ -21,11 +15,11 @@
           <div class="mx-auto text-center">
             <v-avatar color="indigo" style="margin-bottom: 10px">
               <span class="white--text text-h6">{{
-                $store.state.account.initials
+                getInitials(user.account)
               }}</span>
             </v-avatar>
-            <h3>{{ $store.state.account.name }}</h3>
-            <p class="text-caption mt-1">{{ $store.state.account.username }}</p>
+            <h3>{{ user.account.name }}</h3>
+            <p class="text-caption mt-1">{{ user.account.username }}</p>
             <v-divider class="my-3"></v-divider>
             <v-btn @click="logout" depressed text> Disconnect </v-btn>
           </div>
@@ -42,7 +36,7 @@
     >
       <template v-slot:activator="{ on }">
         <v-btn
-          :disabled="!$store.state.msalConfig.auth.authority"
+          :disabled="!user.msalConfig.auth.authority"
           @click="login"
           icon
           x-large
@@ -55,51 +49,43 @@
   </v-col>
 </template>
 
-<script>
+<script setup lang="ts">
 import { mdiLogin } from "@mdi/js";
+import { AccountInfo } from "@azure/msal-browser";
+import { useUserStore } from "../stores/user";
+import { useInternalStore } from "../stores/internal";
+import { sendCommand } from "@/utils";
 
-function getInitials(account) {
+const user = useUserStore();
+const internal = useInternalStore();
+
+function getInitials(account: AccountInfo) {
+  if (account === undefined) return "";
   return account.name.split(" ")[0].substring(0, 1);
 }
 
-export default {
-  data() {
-    return {
-      mdiLogin,
-    };
-  },
-  methods: {
-    async login() {
-      await this.$store.state.msalInstance
-        .loginPopup({ scopes: ["User.Read"] })
-        .then(() => {
-          const accounts = this.$store.state.msalInstance.getAllAccounts();
-          accounts[0].initials = getInitials(accounts[0]);
-          this.$store.commit("setAccount", accounts[0]);
-        })
-        .catch((error) => {
-          console.error(`Error during authentication: ${error}`);
-        });
-      this.$store.commit(
-        "showSnackbar",
-        `Logged in as ${this.$store.state.account.username}`
-      );
+async function login() {
+  await user.msalInstance
+    .loginPopup({ scopes: ["User.Read"] })
+    .then(() => {
+      const accounts = user.msalInstance.getAllAccounts();
+      user.account = accounts[0];
+      internal.showSnackbar(`Logged in as ${user.account.username}`);
       window.location.reload();
-    },
-    async logout() {
-      let subscription =
-        await this.$store.state.sw.pushManager.getSubscription();
+    })
+    .catch((error) => {
+      console.error(`Error during authentication: ${error}`);
+    });
+}
 
-      if (subscription)
-        await this.send_command(
-          `devices?endpoints=${subscription.endpoint}`,
-          "DELETE"
-        );
+async function logout() {
+  let subscription = await internal.sw.pushManager.getSubscription();
 
-      await this.$store.state.msalInstance.logout({}).catch((error) => {
-        console.error(error);
-      });
-    },
-  },
-};
+  if (subscription)
+    await sendCommand(`devices?endpoints=${subscription.endpoint}`, "DELETE");
+
+  await user.msalInstance.logout({}).catch((error) => {
+    console.error(error);
+  });
+}
 </script>
