@@ -11,7 +11,7 @@
           small
           plain
           ><span class="text-truncate" style="max-width: 70px">{{
-            name
+            props.name
           }}</span></v-btn
         >
       </v-container>
@@ -27,7 +27,7 @@
           to you</span
         >
         <v-spacer />
-        <v-text-field v-model="new_name" :placeholder="name" dense />
+        <v-text-field v-model="new_name" :placeholder="props.name" dense />
       </v-card-text>
       <v-card-actions>
         <v-spacer />
@@ -49,17 +49,13 @@
   </v-dialog>
 </template>
 
-<script>
-import { mdiPencil } from "@mdi/js";
+<script setup lang="ts">
+import { ref } from "vue";
 
-export default {
-  props: ["name", "index"],
-  data: function () {
-    return {
-      mdiPencil,
-      new_name: name,
-      dialog: false,
-    };
-  },
-};
+const props = defineProps<{
+  name: string;
+}>();
+
+const new_name = ref(name);
+const dialog = ref(false);
 </script>
