@@ -40,7 +40,7 @@
               class="align-end"
               :color="getPvColor(item, key)"
               text-color="white"
-              :href="`https://${$store.state.url}/archiver-viewer/?pv=${item.pvs[key].name}`"
+              :href="`https://${internal.url}/archiver-viewer/?pv=${item.pvs[key].name}`"
               target="_blank"
             >
               {{ item.pvs[key].value }}
@@ -56,67 +56,56 @@
   </v-card>
 </template>
 
-<script>
+<script setup lang="ts">
 import PowerPanel from "./PowerPanel";
 import ConfigDialog from "./ConfigDialog";
 import NotificationToggler from "./NotificationToggler";
 import { mdiChartAreasplineVariant } from "@mdi/js";
+import { useInternalStore } from "@/stores/internal";
+import Item from "@/models/item";
 
-export default {
-  props: ["item", "filtered_keys"],
-  data: function () {
-    return {
-      dialog: false,
-      mdiChartAreasplineVariant,
-    };
-  },
-  components: {
-    PowerPanel,
-    ConfigDialog,
-    NotificationToggler,
-  },
-  methods: {
-    getPvColor(item, key) {
-      const value = item.pvs[key].value;
+const internal = useInternalStore();
 
-      if (value === "?") return "gray";
+const props = defineProps<{
+  item: Item;
+  filtered_keys: string[];
+}>();
 
-      const m_type = key.charAt(0).toLowerCase();
-      const f_value =
-        m_type !== "h"
-          ? parseFloat(value.substring(0, value.indexOf(" ")))
-          : parseFloat(value.substring(0, value.indexOf("%")));
+function getPvColor(item: Item, key: string) {
+  const value = item.pvs[key].value;
 
-      if (key === "Rack Open" || key == "Leak") {
-        return value === "No" ? "green" : "orange";
-      }
+  if (value === "?") return "gray";
 
-      if (
-        f_value > item.pvs[key].hi_limit * 1.2 ||
-        f_value < item.pvs[key].lo_limit * 0.8
-      )
-        return "red";
-      else if (
-        f_value > item.pvs[key].hi_limit ||
-        f_value < item.pvs[key].lo_limit
-      )
-        return "orange";
-      else return "green";
-    },
-    display_archiver() {
-      let pv_names = [];
-      for(const pv of Object.values(this.item.pvs)) {
-        if (pv.name !== "")
-          pv_names.push(pv.name);
-      }
+  const m_type = key.charAt(0).toLowerCase();
+  const f_value =
+    m_type !== "h"
+      ? parseFloat(value.substring(0, value.indexOf(" ")))
+      : parseFloat(value.substring(0, value.indexOf("%")));
 
-      window.open(
-        `https://${
-          this.$store.state.url
-        }/archiver-viewer/?pv=${pv_names.join("&pv=")}`,
-        "_blank"
-      );
-    },
-  },
-};
+  if (key === "Rack Open" || key == "Leak") {
+    return value === "No" ? "green" : "orange";
+  }
+
+  if (
+    f_value > item.pvs[key].hi_limit * 1.2 ||
+    f_value < item.pvs[key].lo_limit * 0.8
+  )
+    return "red";
+  else if (f_value > item.pvs[key].hi_limit || f_value < item.pvs[key].lo_limit)
+    return "orange";
+  else return "green";
+}
+function display_archiver() {
+  let pv_names = [];
+  for (const pv of Object.values(props.item.pvs)) {
+    if (pv.name !== "") pv_names.push(pv.name);
+  }
+
+  window.open(
+    `https://${internal.url}/archiver-viewer/?pv=${pv_names.join("&pv=")}`,
+    "_blank"
+  );
+}
+
+defineExpose({getPvColor});
 </script>
