@@ -2,43 +2,34 @@ import Vuetify from 'vuetify'
 import { createLocalVue, mount } from '@vue/test-utils'
 import RackCard from '@/components/RackCard.vue'
 import { EMPTY_PVS } from '@/assets/constants'
-import Vuex from 'vuex'
-
+import { createTestingPinia } from '@pinia/testing'
+import { useUserStore } from '@/stores/user'
+import { defineStore, PiniaVuePlugin } from 'pinia'
 
 describe('RackCard.vue', () => {
   const keys = ["Temperature", "Pressure", "Rack Open", "Humidity", "Leak"]
   const pvs = Object.assign({}, EMPTY_PVS, {Temperature: {name: "960E1:CO-SIMAR-01:Temp-Mon", value: "20 C"}, Pressure: {name: "960E1:CO-SIMAR-01:Pressure-Mon", value: "900 hPa"}});
 
   const default_item = { name: "Test", parent: "Test", pv_names: ["960E1:CO-SIMAR-01:Temp-Mon", "960E1:CO-SIMAR-01:Pressure-Mon"], pvs: pvs };
-  const localVue = createLocalVue()
-  localVue.use(Vuex);
-  let vuetify
-  let store
+  const localVue = createLocalVue();
+  localVue.use(PiniaVuePlugin);
+  let vuetify;
+  let store;
 
   beforeEach(() => {
     vuetify = new Vuetify()
-
-    store = new Vuex.Store({
-      state: {
-        account: undefined,
-        url: "ais-eng-srv-la.cnpem.br",
-      },
-      mutations: {
-        setAccount(state, account) {
-          state.account = account;
-        },
-      }
-    })
   })
 
   const mountFunction = options => {
     return mount(RackCard, {
       localVue,
       vuetify,
-      store,
-      propsData: {item: default_item, filtered_keys: keys},
+      pinia: createTestingPinia({initialState: {user: {account: undefined}, internal: {url: "pudim.com.br"}}}),
+      propsData: { item: default_item, filtered_keys: keys},
       mocks: {
         $vuetify: { breakpoint: {} },
+        item: default_item,
+        filtered_keys: keys
       },
       ...options
     })
@@ -64,10 +55,11 @@ describe('RackCard.vue', () => {
   })
 
   it('modifies PV values in real time', async () => {
-    let item = default_item;
+    const item = default_item;
     const wrapper = mountFunction();
 
     expect(wrapper.find(".v-card").find(".v-list").findAll("span .v-chip__content").at(0).text()).toMatch("20 C");
+    console.log(wrapper.html());
 
     item.pvs.Temperature.value = "22 C";
     await wrapper.setProps(item);
@@ -83,7 +75,7 @@ describe('RackCard.vue', () => {
   })
 
   it('enables configuration for authenticated users', () => {
-    store.commit("setAccount", true);
+    store.account = true;
     const wrapper = mountFunction();
     const button = wrapper.find(".v-card").find(".v-card__title").find(".v-dialog__container").find(".v-btn--disabled");
 
@@ -91,7 +83,7 @@ describe('RackCard.vue', () => {
   })
 
   it('sets chip links to valid Archiver URLs', async () => {
-    let item = default_item;
+    const item = default_item;
     const wrapper = mountFunction();
 
     item.pvs.Temperature.name = "960E1:CO-SIMAR-01:Temp-Mon";
